+++ cynic/src/scalar.rs
@@ -77,3 +77,77 @@ where
         S::decode(value)
     }
 }
+
+#[cfg(feature = "chrono")]
+impl Scalar for chrono::DateTime<chrono::FixedOffset> {
+    fn decode(value: &serde_json::Value) -> Result<Self, DecodeError> {
+        use chrono::DateTime;
+
+        match value {
+            serde_json::Value::String(s) => {
+                Ok(DateTime::parse_from_rfc3339(s).map_err(chrono_decode_error)?)
+            }
+            _ => Err(DecodeError::IncorrectType(
+                "String".to_string(),
+                value.to_string(),
+            )),
+        }
+    }
+
+    fn encode(&self) -> Result<serde_json::Value, ()> {
+        Ok(serde_json::Value::String(self.to_rfc3339()))
+    }
+}
+
+#[cfg(feature = "chrono")]
+impl Scalar for chrono::DateTime<chrono::Utc> {
+    fn decode(value: &serde_json::Value) -> Result<Self, DecodeError> {
+        use chrono::{DateTime, Utc};
+
+        match value {
+            serde_json::Value::String(s) => Ok(DateTime::parse_from_rfc3339(s)
+                .map_err(chrono_decode_error)?
+                .with_timezone(&Utc)),
+            _ => Err(DecodeError::IncorrectType(
+                "String".to_string(),
+                value.to_string(),
+            )),
+        }
+    }
+
+    fn encode(&self) -> Result<serde_json::Value, ()> {
+        Ok(serde_json::Value::String(self.to_rfc3339()))
+    }
+}
+
+#[cfg(feature = "chrono")]
+fn chrono_decode_error(err: chrono::ParseError) -> DecodeError {
+    DecodeError::Other(err.to_string())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[cfg(feature = "chrono")]
+    #[test]
+    fn test_utc_datetime_scalar() {
+        use chrono::{DateTime, Utc};
+
+        let datetime: DateTime<Utc> = Utc::now();
+
+        assert_eq!(DateTime::decode(&datetime.encode().unwrap()), Ok(datetime));
+    }
+
+    #[cfg(feature = "chrono")]
+    #[test]
+    fn test_fixed_offset_datetime_scalar() {
+        use chrono::{DateTime, FixedOffset, TimeZone};
+
+        let datetime: DateTime<FixedOffset> = FixedOffset::east(3600 * 5)
+            .ymd(2016, 11, 08)
+            .and_hms(10, 15, 20);
+
+        assert_eq!(DateTime::decode(&datetime.encode().unwrap()), Ok(datetime));
+    }
+}